requirements.txt
psycopg2-binary>=2.9
python-dotenv>=0.21
requests>=2.28
pandas>=1.5
fastapi>=0.100.0
uvicorn>=0.23.0
